Fixed String updating & moving code for Integer conversion to Recursive Function
Fixed the issue of not being able to update the string, by simply declaring the string again instead of trying to update it (which is not allowed) https://stackoverflow.com/questions/11924689/change-string-in-java
https://stackoverflow.com/questions/11164675/mutable-strings-in-java

Moved the code for converting the user input to an integer to its own method, so I can refer to itself - changing it into a recursive function.

Got a new issue with the String not being convertable to a variable. Think it's an easy fix but I am out of time right now. Will resume later.